requirements.txt
torch==2.3.0
torchvision==0.18.0
torchaudio==2.3.0
streamlit>=1.29.0
fastapi>=0.104.0
uvicorn>=0.24.0
transformers>=4.35.0
huggingface-hub>=0.19.0
graphviz>=0.20.1
matplotlib>=3.7.0
seaborn>=0.13.0
plotly>=5.18.0
pandas>=2.0.0
numpy>=1.21.0
jax>=0.4.0
jaxlib>=0.4.0
fpdf>=1.7.2
Pillow>=9.0.0
tqdm
jinja2